pyproject.toml
[project]
name = "sales-chat-ingestion"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "duckdb==1.1.3",
  "pydantic==2.10.3",
  "python-multipart==0.0.18",
  "pandas==2.2.3",
  "openpyxl==3.1.5",
  "google-generativeai==0.8.3",
  "langgraph==0.2.56",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
